refactor: components 폴더의 interview 도메인 구조 정리
