+++ src/components/interview/InterviewingView/index.tsx
@@ -1,0 +1,3 @@
+export function InterviewingView() {
+  return;
+}
